all error and bugs fixed

+++ lib/controller/noteController.dart
@@ -1,7 +1,7 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:note_app/controller/authController.dart';
 import 'package:note_app/model/model.dart';
 
@@ -37,43 +37,51 @@ class NoteController extends GetxController {
   NoteModel noteModel = Get.put(NoteModel());
 
   void addNote() async {
-    var notemodel = NoteModel(
-      note: addnote.text,
-      des: adddes.text,
-      userName: authController.userName.text,
-      noteI: DateTime.now().millisecondsSinceEpoch.toString(),
-    );
-    await db
-        .collection("users")
-        .doc(auth.currentUser!.uid)
-        .collection("note")
-        .doc(notemodel.noteI)
-        .set(notemodel.toJson());
+    var currentUser = auth.currentUser;
+    if (currentUser != null) {
+      var notemodel = NoteModel(
+        note: addnote.text,
+        des: adddes.text,
+        userName: authController.userName.text,
+        noteI: DateTime.now().millisecondsSinceEpoch.toString(),
+      );
+      await db
+          .collection("users")
+          .doc(currentUser.uid)
+          .collection("note")
+          .doc(notemodel.noteI)
+          .set(
+            notemodel.toJson(),
+          );
 
-    getNote();
-    addnote.clear();
-    adddes.clear();
+      getNote();
+      addnote.clear();
+      adddes.clear();
 
-    Get.snackbar("Note Added", "Note Added to Firestore",
-        backgroundColor: Colors.green);
+      Get.snackbar("Note Added", "Note Added to Firestore",
+          backgroundColor: Colors.green);
+    }
   }
 
   Future<void> getNote() async {
     try {
-      var data = await db
-          .collection("users")
-          .doc(auth.currentUser!.uid)
-          .collection("note")
-          .get();
-      noteList.clear();
+      var currentUser = auth.currentUser;
+      if (currentUser != null) {
+        var data = await db
+            .collection("users")
+            .doc(currentUser.uid)
+            .collection("note")
+            .get();
+        noteList.clear();
 
-      for (var note in data.docs) {
-        noteList.add(NoteModel.fromJson(note.data()));
+        for (var note in data.docs) {
+          noteList.add(NoteModel.fromJson(note.data()));
+        }
+
+        hasNote.value = noteList.isNotEmpty; // Update hasNote based on noteList
+
+        noteList.refresh();
       }
-
-      hasNote.value = noteList.isNotEmpty; // Update hasNote based on noteList
-
-      noteList.refresh();
     } catch (ex) {
       Get.snackbar("Error", ex.toString());
     }
@@ -81,35 +89,41 @@ class NoteController extends GetxController {
 
   Future<void> deleteNote(String noteId) async {
     try {
-      await db
-          .collection("users")
-          .doc(auth.currentUser!.uid)
-          .collection("note")
-          .doc(noteId)
-          .delete()
-          .then((value) => {
-                Get.snackbar(
-                    "Note Deleted", "Note Successfully Deleted form Firestore",
-                    backgroundColor: Colors.green)
-              });
-      getNote();
+      var currentUser = auth.currentUser;
+      if (currentUser != null) {
+        await db
+            .collection("users")
+            .doc(currentUser.uid)
+            .collection("note")
+            .doc(noteId)
+            .delete()
+            .then((value) => {
+                  Get.snackbar("Note Deleted",
+                      "Note Successfully Deleted form Firestore",
+                      backgroundColor: Colors.green)
+                });
+        getNote();
+      }
     } catch (ex) {
       Get.snackbar("Error", ex.toString());
     }
   }
 
   Future<void> editNote(String noteDocID) async {
-    await db
-        .collection("users")
-        .doc(auth.currentUser!.uid)
-        .collection("note")
-        .doc(noteDocID)
-        .update({
-      "note": editnote.text,
-      "des": editdes.text,
-    });
-    editnote.clear();
-    editdes.clear();
-    getNote();
+    var currentUser = auth.currentUser;
+    if (currentUser != null) {
+      await db
+          .collection("users")
+          .doc(currentUser.uid)
+          .collection("note")
+          .doc(noteDocID)
+          .update({
+        "note": editnote.text,
+        "des": editdes.text,
+      });
+      editnote.clear();
+      editdes.clear();
+      getNote();
+    }
   }
 }
